+++ Cudheart/Testing/Math/Exps/TestExps.cu
@@ -1,0 +1,8 @@
+#include "TestExps.cuh"
+
+
+namespace Cudheart::Testing::Math {
+	void testExps() {
+		CPP::Exps::testExps();
+	}
+}